Implement a basic iroh-automerge-repo library
